Refactor response logic and reorganize vote endpoint
